pyproject.toml
[project]
name = "lsj-quiz"
version = "0.1.0"
description = ""
authors = [
    {name = "soojiinleee",email = "soojiin.leee@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "shortuuid (>=1.0.13,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "pytest-django (>=4.11.0,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "lsj_quiz", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
